requirements.txt
streamlit
pytest
google-genai
python-dotenv
